Light refactoring for UserFactory
After updating my branch, there were some errors with imports as a result of some classes being in the wrong place. I moved them into the proper packages, then made sure the tests still worked after their import statements were appropriately adjusted.
